+++ tests/test_ckan_backend_filters.py
@@ -30,7 +30,7 @@ def package_factory(ckan, **kwargs):
     return response['result']
 
 
-def test_org_filter(ckan):
+def test_include_org_filter(ckan):
     # create 2 organizations with 2 datasets each
     org = ckan.action('organization_create', {'name': 'org-1'})['result']
     included_ids = [d['id'] for d in [
@@ -55,6 +55,34 @@ def test_org_filter(ckan):
         assert dataset.extras['harvest:remote_id'] in included_ids
 
 
+def test_exclude_org_filter(ckan):
+    # create 2 organizations with 2 datasets each
+    org = ckan.action('organization_create', {'name': 'org-1'})['result']
+    included_ids = [d['id'] for d in [
+        package_factory(ckan, owner_org=org['id']),
+        package_factory(ckan, owner_org=org['id']),
+    ]]
+    org2 = ckan.action('organization_create', {'name': 'org-2'})['result']
+    excluded_ids = [d['id'] for d in [
+        package_factory(ckan, owner_org=org2['id']),
+        package_factory(ckan, owner_org=org2['id']),
+    ]]
+
+    source = HarvestSourceFactory(backend='ckan', url=ckan.BASE_URL, config={
+        'filters': [{'key': 'organization', 'value': org2['name'], 'type': 'exclude'}]
+    })
+
+    actions.run(source.slug)
+    source.reload()
+
+    job = source.get_last_job()
+    assert len(job.items) == len(included_ids)
+
+    for dataset in Dataset.objects:
+        assert dataset.extras['harvest:remote_id'] in included_ids
+        assert dataset.extras['harvest:remote_id'] not in excluded_ids
+
+
 def test_tag_filter(ckan):
     # create 2 datasets with a different tag each
     tag = faker.word()
@@ -74,6 +102,25 @@ def test_tag_filter(ckan):
     assert Dataset.objects.first().extras['harvest:remote_id'] == package['id']
 
 
+def test_exclude_tag_filter(ckan):
+    # create 2 datasets with a different tag each
+    tag = faker.word()
+    package_factory(ckan, tags=[{'name': tag}])
+    included = package_factory(ckan, tags=[{'name': faker.word()}])
+
+    source = HarvestSourceFactory(backend='ckan', url=ckan.BASE_URL, config={
+        'filters': [{'key': 'tags', 'value': tag, 'type': 'exclude'}]
+    })
+
+    actions.run(source.slug)
+    source.reload()
+
+    job = source.get_last_job()
+    assert len(job.items) == 1
+    assert Dataset.objects.count() == 1
+    assert Dataset.objects.first().extras['harvest:remote_id'] == included['id']
+
+
 def test_can_have_multiple_filters(ckan):
     # create 2 organizations with 2 datasets each
     org = ckan.action('organization_create', {'name': 'org-1'})['result']
@@ -86,7 +133,7 @@ def test_can_have_multiple_filters(ckan):
     source = HarvestSourceFactory(backend='ckan', url=ckan.BASE_URL, config={
         'filters': [
             {'key': 'organization', 'value': org['name']},
-            {'key': 'tags', 'value': 'tag-1'},
+            {'key': 'tags', 'value': 'tag-2', 'type': 'exclude'},
         ]
     })