+++ udata_ckan/harvesters.py
@@ -6,7 +6,7 @@ from uuid import UUID
 from urllib.parse import urljoin
 
 from voluptuous import (
-    Schema, All, Any, Lower, Coerce, DefaultTo
+    Schema, All, Any, Lower, Coerce, DefaultTo, Optional
 )
 
 from udata import uris
@@ -88,7 +88,7 @@ schema = Schema({
     'organization': Any(organization, None),
     'resources': [resource],
     'revision_id': str,
-    'extras': [{
+    Optional('extras', default=list): [{
         'key': str,
         'value': Any(str, int, float, boolean, dict, list),
     }],
@@ -131,10 +131,35 @@ class CkanBackend(BaseBackend):
             response = self.post(url, '{}', params=kwargs)
         else:
             response = self.get(url, params=kwargs)
-        if response.status_code != 200:
+
+        content_type = response.headers.get('Content-Type', '')
+        mime_type = content_type.split(';', 1)[0]
+
+        if mime_type == 'application/json':  # Standard API JSON response
+            data = response.json()
+            # CKAN API always returns 200 even on errors
+            # Only the `success` property allows to detect errors
+            if data.get('success', False):
+                return data
+            else:
+                error = data.get('error')
+                if isinstance(error, dict):
+                    # Error object with message
+                    msg = error.get('message', 'Unknown error')
+                    if '__type' in error:
+                        # Typed error
+                        msg = ': '.join((error['__type'], msg))
+                else:
+                    # Error only contains a message
+                    msg = error
+                raise HarvestException(msg)
+
+        elif mime_type == 'text/html':  # Standard html error page
+            raise HarvestException('Unknown Error: {} returned HTML'.format(url))
+        else:
+            # If it's not HTML, CKAN respond with raw quoted text
             msg = response.text.strip('"')
             raise HarvestException(msg)
-        return response.json()
 
     def get_status(self):
         url = urljoin(self.source.url, '/api/util/status')
@@ -152,7 +177,10 @@ class CkanBackend(BaseBackend):
             # use q parameters because fq is broken with multiple filters
             params = []
             for f in filters:
-                params.append('{key}:{value}'.format(**f))
+                param = '{key}:{value}'.format(**f)
+                if f.get('type') == 'exclude':
+                    param = '-' + param
+                params.append(param)
             q = ' AND '.join(params)
             response = self.get_action('package_search', fix=fix, q=q)
             names = [r['name'] for r in response['result']['results']]